feat(email): migrate digest onto shared yas.email base + macros
Moves DigestPayload, gather_digest_payload, and digest templates into the
yas.email package; introduces base.{html,txt}.j2 + macros.j2 for reuse by
the upcoming immediate-alert renderers. yas.alerts.digest.builder and
yas.alerts.digest.filters are now thin re-export shims; existing callers
(worker/digest_loop.py, web/routes/digest_preview.py, tests) keep working
unchanged.

New public API in yas.email:
  - render_email(session, kind, lead, members) -> RenderedEmail
    Looks up a TypeRenderer in RENDERERS and renders its pair.
  - render_digest_payload(payload, top_line) -> RenderedEmail
    Sync entry point used by the digest worker which assembles its own
    payload outside the Alert lifecycle. Looks up RENDERERS[AlertType.digest]
    directly (cleaner than the dual-key fallback the plan sketched).
  - Subject is pulled out of {% block subject %} in the .txt template via
    txt_tpl.blocks["subject"], so each template owns its own subject.

Golden re-baseline (chrome + macro/whitespace):
  - DOCTYPE/body style is the shared base's, header/footer replace the old
    inline <p>/<hr> chrome.
  - The HTML NEW MATCHES <li> now comes from offering_row_html in macros.j2;
    its inner whitespace differs from the old inline <li> (the macro emits a
    single tightly-packed line where the old template had each &middot;
    fragment on its own indented line). Semantic content is unchanged: same
    offering name, site, rel_date, price, score, and Register link in the
    same order. In a real mail client this renders identically because HTML
    collapses whitespace.
  - The TXT NEW MATCHES row is inlined (not via offering_row_text) on
    purpose: the plan's offering_row_text macro appends the registration_url
    to the text row, which would be new content the pre-migration digest
    never emitted. Inlining preserves byte-identical section content;
    offering_row_text stays defined in macros.j2 for future immediate-alert
    templates where the URL is desirable.
  - trim_blocks=True/lstrip_blocks=True (set on the shared env) collapse
    the extra blank lines the old non-trim env left between sections. This
    is whitespace only.

Env note: yas.email.environment has autoescape on for *.html.j2, so HTML
output now escapes apostrophes (' -> &#39;). One assertion in
test_render_digest_includes_top_line was updated to expect the escaped
form; the plain-text branch is unaffected.

The xfail in tests/unit/test_email_registry.py stays xfail; Task 14 flips
it green once every AlertType has a renderer.

Refs spec: docs/superpowers/specs/2026-05-19-outbound-email-template-layer-design.md
