+++ src/yas/email/environment.py
@@ -11,13 +11,9 @@ from pathlib import Path
 
 from jinja2 import Environment, FileSystemLoader, StrictUndefined, select_autoescape
 
-# Temporary import: filters currently live in yas.alerts.digest. Task 4 of the
-# email-template-layer plan relocates them to yas.email.filters and replaces
-# yas.alerts.digest.filters with a re-export shim. Do not add new
-# yas.email -> yas.alerts imports — this is the only one.
-from yas.alerts.digest.filters import fmt as fmt_filter
-from yas.alerts.digest.filters import price as price_filter
-from yas.alerts.digest.filters import rel_date as rel_date_filter
+from yas.email.filters import fmt as fmt_filter
+from yas.email.filters import price as price_filter
+from yas.email.filters import rel_date as rel_date_filter
 
 _TEMPLATES_DIR = Path(__file__).parent / "templates"
