+++ src/yas/email/templates/base.txt.j2
@@ -1,0 +1,6 @@
+{% block subject %}YAS — {{ payload.kid_name | default('Update') }}{% endblock %}
+
+{% block body %}{% endblock %}
+
+---
+Youth Activity Scheduler
