+++ src/yas/email/templates/digest.html.j2
@@ -1,11 +1,8 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <title>Activity Digest — {{ payload.for_date }}</title>
-</head>
-<body style="font-family: sans-serif; max-width: 600px; margin: 0 auto; padding: 16px;">
-
+{% extends "base.html.j2" %}
+{% from "macros.j2" import offering_row_html %}
+{% block title %}Activity Digest — {{ payload.for_date }}{% endblock %}
+{% block heading %}Daily digest — {{ payload.kid_name }}{% endblock %}
+{% block body %}
 <p style="font-size: 1.1em; font-weight: bold;">{{ top_line }}</p>
 
 {% if payload.under_no_matches_threshold %}
@@ -17,16 +14,7 @@ We're still scanning sites and will be in touch as soon as we find something. Ba
 <h1 style="font-size: 1em; border-bottom: 1px solid #ccc;">New Matches ({{ payload.new_matches|length }})</h1>
 <ul>
   {% for m in payload.new_matches %}
-  <li>
-    <strong>{{ m.offering_name }}</strong>
-    {% if m.site_name %} @ {{ m.site_name }}{% endif %}
-    &middot; {{ m.start_date|rel_date }}
-    &middot; {{ m.price_cents|price }}
-    &middot; score {{ "%.2f"|format(m.score) }}
-    {% if m.registration_url %}
-    &middot; <a href="{{ m.registration_url }}">Register</a>
-    {% endif %}
-  </li>
+  {{ offering_row_html(m) }}
   {% endfor %}
 </ul>
 {% endif %}
@@ -63,8 +51,4 @@ We're still scanning sites and will be in touch as soon as we find something. Ba
 {% endif %}
 
 {% endif %}
-
-<hr style="margin-top: 32px;">
-<p style="font-size: 0.8em; color: #999;">Youth Activity Scheduler &middot; {{ payload.for_date }}</p>
-</body>
-</html>
+{% endblock %}
