+++ src/yas/email/templates/macros.j2
@@ -1,0 +1,23 @@
+{# Reusable fragments shared by digest and immediate alerts. #}
+
+{% macro offering_row_html(m) %}
+<li>
+  <strong>{{ m.offering_name }}</strong>
+  {% if m.site_name %} @ {{ m.site_name }}{% endif %}
+  {% if m.start_date %}&middot; {{ m.start_date | rel_date }}{% endif %}
+  {% if m.price_cents is not none %}&middot; {{ m.price_cents | price }}{% endif %}
+  {% if m.score is defined %}&middot; score {{ "%.2f" | format(m.score) }}{% endif %}
+  {% if m.registration_url %}&middot; <a href="{{ m.registration_url }}">Register</a>{% endif %}
+</li>
+{% endmacro %}
+
+{% macro offering_row_text(m) -%}
+  - {{ m.offering_name }}{% if m.site_name %} @ {{ m.site_name }}{% endif %}{% if m.start_date %} · {{ m.start_date | rel_date }}{% endif %}{% if m.price_cents is not none %} · {{ m.price_cents | price }}{% endif %}{% if m.score is defined %} · score {{ "%.2f" | format(m.score) }}{% endif %}{% if m.registration_url %} · {{ m.registration_url }}{% endif %}
+{%- endmacro %}
+
+{% macro reg_countdown_html(opens_at, label) %}
+<div style="background: #fff8e1; border-left: 4px solid #f0a500; padding: 12px; margin: 16px 0;">
+  <strong>{{ label }}</strong>
+  <div style="color: #555; font-size: 0.9em; margin-top: 4px;">{{ opens_at | fmt }}</div>
+</div>
+{% endmacro %}
