+++ tests/golden/digest/empty.html
@@ -4,15 +4,12 @@
   <meta charset="utf-8">
   <title>Activity Digest — 2026-05-19</title>
 </head>
-<body style="font-family: sans-serif; max-width: 600px; margin: 0 auto; padding: 16px;">
+<body style="font-family: -apple-system, Segoe UI, sans-serif; max-width: 600px; margin: 0 auto; padding: 16px; color: #222;">
+<header style="border-bottom: 1px solid #ccc; padding-bottom: 8px; margin-bottom: 16px;">
+  <strong style="font-size: 1.05em;">Daily digest — Bo</strong>
+</header>
 
 <p style="font-size: 1.1em; font-weight: bold;">Bo — quiet day</p>
-
-
-
-
-
-
 
 
 
@@ -22,9 +19,7 @@
 <p style="color: #666;">No new activity today.</p>
 
 
-
-
-<hr style="margin-top: 32px;">
-<p style="font-size: 0.8em; color: #999;">Youth Activity Scheduler &middot; 2026-05-19</p>
+<footer style="margin-top: 32px; padding-top: 8px; border-top: 1px solid #eee; font-size: 0.8em; color: #999;">
+  Youth Activity Scheduler · 2026-05-19</footer>
 </body>
 </html>
