+++ tests/golden/digest/empty.txt
@@ -1,7 +1,4 @@
 Bo — quiet day
-
-
-
-
-
 No new activity today.
+---
+Youth Activity Scheduler
