+++ tests/golden/digest/under_threshold.txt
@@ -1,3 +1,4 @@
 Cy — still searching
-
 We haven't found any activities matching Cy yet. We're still scanning sites and will be in touch as soon as we find something. Back tomorrow.
+---
+Youth Activity Scheduler
