+++ tests/golden/digest/with_matches.html
@@ -4,27 +4,20 @@
   <meta charset="utf-8">
   <title>Activity Digest — 2026-05-19</title>
 </head>
-<body style="font-family: sans-serif; max-width: 600px; margin: 0 auto; padding: 16px;">
+<body style="font-family: -apple-system, Segoe UI, sans-serif; max-width: 600px; margin: 0 auto; padding: 16px; color: #222;">
+<header style="border-bottom: 1px solid #ccc; padding-bottom: 8px; margin-bottom: 16px;">
+  <strong style="font-size: 1.05em;">Daily digest — Ada</strong>
+</header>
 
 <p style="font-size: 1.1em; font-weight: bold;">Ada — 1 new match</p>
 
 
-
-
 <h1 style="font-size: 1em; border-bottom: 1px solid #ccc;">New Matches (1)</h1>
 <ul>
-  
   <li>
-    <strong>Soccer Camp</strong>
-     @ Park District
-    &middot; Mon, Jun 1
-    &middot; $150.00
-    &middot; score 0.91
-    
-    &middot; <a href="https://example.com/reg/10">Register</a>
-    
-  </li>
-  
+  <strong>Soccer Camp</strong>
+ @ Park District&middot; Mon, Jun 1&middot; $150.00&middot; score 0.91&middot; <a href="https://example.com/reg/10">Register</a></li>
+
 </ul>
 
 
@@ -32,13 +25,7 @@
 
 
 
-
-
-
-
-
-
-<hr style="margin-top: 32px;">
-<p style="font-size: 0.8em; color: #999;">Youth Activity Scheduler &middot; 2026-05-19</p>
+<footer style="margin-top: 32px; padding-top: 8px; border-top: 1px solid #eee; font-size: 0.8em; color: #999;">
+  Youth Activity Scheduler · 2026-05-19</footer>
 </body>
 </html>
