+++ tests/golden/digest/with_matches.txt
@@ -1,8 +1,6 @@
 Ada — 1 new match
-
 NEW MATCHES (1):
 - Soccer Camp @ Park District · Mon, Jun 1 · $150.00 · score 0.91
 
-
-
-
+---
+Youth Activity Scheduler
